Refactor: 팀원 작업 내용 반영 후 구조 정리
